Source code update

+++ Stage/PlayerExp.cs
@@ -5,7 +5,7 @@ public class PlayerExp : MonoBehaviour
 {
     private void Update()
     {
-        GetComponent<Slider>().value = (float)PlayerPrefs.GetInt("XP")
-                                     / (float)PlayerPrefs.GetInt("LV");
+        GetComponent<Slider>().value = PlayerPrefs.GetFloat("XP")
+                                     / PlayerPrefs.GetInt("LV");
     }
 }
